Added tests dealing with attributes.
It seems that System.Xml.Serialization.XmlSerializer doesn't like it if
you decorate a derived class' property with a different attribute than
what the base class' property is decorated with. CustomSerializer should
throw an exception under the same circumstances that the BCL
XmlSerializer does, and *not* throw an exception when the BCL
XmlSerializer doesn't.

Note that many of these tests are failing as of this commit.