Added ability to serialize dynamic properties.
This commit only addresses serialization - deserialization has not yet
been implemented.
ExpandoObject serialization has not yet been implemented.
Serializing for an object of type object is now done exactly the same as
an object of type dynamic.
The keys and values of non-generic dictionaries are now serialized the
same as dynamic and object.
Added option to always emit the xsd type when serializing an instance of
type object or dynamic.