Remove implicit conversions to Lists, improve dynamic conversions
The result of a dynamic conversion (to IEnumerable<T>, ICollection<T> or IList<T>) is an implementation
of IList<T> which has a reference to the JsonArray's `List<object> _values`. It uses that list in order
to modify the original collection during write operations (Add, Insert, Delete, etc.) of the custom
IList<T>.

Support for implicit conversion to List<T> was removed because they were "lied" to users - when the
resulting list was modified, the JsonArray did not change.