JsonObject: override equality members, add gaurd to add/set members
These changes make JsonObject suitable to be the key in a dictionary. Values in the _values
dictionary will only ever be one of: null, bool, string, JsonNumber, JsonObject, or JsonArray.
However, if a Guid, DateTime, or DateTimeOffset is used as the value for a JsonObject, the
value will be converted to a string before storing in the _values dictionary. Similarly, if
a numeric value (int, double, etc) is used, it will be converted to a JsonNumber object.