XSerializer version 0.4.1
When deserializing into a class that has properties marked with [Encrypt], handle (and test for) the case where the property value is not actually encrypted. Whether a value is encrypted or plain-text is determined implicitly by the encryption mechanism. If its Decrypt method returns the same value passed to it, the value is assumed to be plain-text.